Refactoring: move code in to methods and classes in to files
